カードリストをコンポーネント化し、PageViewにする
